+++ src/main/java/me/pajic/tiered_backpacks/tooltip/BackpackTooltipCompat.java
@@ -5,6 +5,7 @@ import com.misterpemodder.shulkerboxtooltip.api.provider.PreviewProviderRegistry
 import me.pajic.tiered_backpacks.TieredBackpacks;
 import me.pajic.tiered_backpacks.util.BackpackUtil;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Item;
 import org.jetbrains.annotations.NotNull;
 
@@ -15,8 +16,9 @@ public class BackpackTooltipCompat implements ShulkerBoxTooltipApi {
 
     @Override
     public void registerProviders(@NotNull PreviewProviderRegistry registry) {
-        Set<Item> backpacks = new HashSet<>();
-        BuiltInRegistries.ITEM.getTagOrEmpty(BackpackUtil.BACKPACKS).forEach(item -> backpacks.add(item.value()));
-        registry.register(TieredBackpacks.id("backpack_tooltip"), new BackpackPreviewProvider(), backpacks);
+        Set<Item> items = new HashSet<>();
+        BuiltInRegistries.ITEM.getTagOrEmpty(BackpackUtil.BACKPACKS).forEach(item -> items.add(item.value()));
+		BuiltInRegistries.ITEM.getTagOrEmpty(ItemTags.CHEST_ARMOR).forEach(item -> items.add(item.value()));
+        registry.register(TieredBackpacks.id("backpack_tooltip"), new BackpackPreviewProvider(), items);
     }
 }